Updated scripts for automation of benchmarking.
`get_kern.py` tracks the kernel versions of the test machines. It calls `ls /boot` and parses the output by lines. However, on my machine, there is an item called `vmlinuz.old`, which causes the script to fail.
Since the script takes care of versions starting with `vmlinuz-`, I modified the script to ignore `vmlinuz.old`.

`run.py` is still for python 2, I transformed it to python 3 by `2to3 -w run.py`. Since we'd like to automate the benchmarking process, I also added a few lines to allow customization of the directory of input and output.